+++ xclim/testing/tests/test_run_length.py
@@ -36,9 +36,9 @@ class TestRLE:
             da = da.chunk({"a": 1, "b": 2})
 
         out = rl.rle(da != 0).mean(["a", "b", "c"])
-        expected = np.zeros(366)
+        expected = np.zeros(365)
         expected[1] = 10
-        expected[2:12] = np.nan
+        expected[2:11] = np.nan
         np.testing.assert_array_equal(out, expected)
 
 
@@ -245,6 +245,54 @@ class TestLastRun:
 
         out = rl.last_run(runs, window=1, dim="time", coord=coord, ufunc_1dim=use_1dim)
         np.testing.assert_array_equal(out.load(), expected)
+
+
+def test_run_bounds_synthetic():
+    run = xr.DataArray(
+        [0, 1, 1, 1, 0, 0, 1, 1, 1, 0], dims="x", coords={"x": np.arange(10) ** 2}
+    )
+    bounds = rl.run_bounds(run, "x", coord=True)
+    np.testing.assert_array_equal(bounds, [[1, 36], [16, 81]])
+
+    bounds = rl.run_bounds(run, "x", coord=False)
+    np.testing.assert_array_equal(bounds, [[1, 6], [4, 9]])
+
+
+def test_run_bounds_data():
+    era5 = open_dataset("ERA5/daily_surface_cancities_1990-1993.nc")
+    cond = era5.tas.rolling(time=7).mean() > 285
+
+    bounds = rl.run_bounds(cond, "time")  # def coord = True
+    np.testing.assert_array_equal(
+        bounds.isel(location=0, events=0),
+        pd.to_datetime(["1990-06-19", "1990-10-26"]).values,
+    )
+
+    bounds = rl.run_bounds(cond, "time", coord="dayofyear")
+    np.testing.assert_array_equal(bounds.isel(location=1, events=4), [279, 283])
+    assert bounds.events.size == 15
+
+
+def test_keep_longest_run_synthetic():
+    runs = xr.DataArray([0, 1, 1, 1, 0, 0, 1, 1, 1, 0], dims="x").astype(bool)
+    lrun = rl.keep_longest_run(runs, "x")
+    np.testing.assert_array_equal(
+        lrun, np.array([0, 1, 1, 1, 0, 0, 0, 0, 0, 0], dtype=bool)
+    )
+
+
+def test_keep_longest_run_data():
+    era5 = open_dataset("ERA5/daily_surface_cancities_1990-1993.nc")
+    cond = era5.snd > 0.001
+    lrun = rl.keep_longest_run(cond, "time")
+    np.testing.assert_array_equal(
+        lrun.isel(time=slice(651, 658), location=2),
+        np.array([0, 0, 0, 1, 1, 1, 1], dtype=bool),
+    )
+
+    xr.testing.assert_equal(
+        rl.keep_longest_run(cond, "time").sum("time"), rl.longest_run(cond, "time")
+    )
 
 
 class TestRunsWithDates: