+++ HISTORY.rst
@@ -8,12 +8,18 @@ History
 New indicators
 ~~~~~~~~~~~~~~
 * `days_with_snow`, which allows counting days with snow between low and high thresholds, e.g. days with high amount of snow.
+* `fire_season`, create a fire season mask from temperature and, optionnally, snow depth timeseries.
 
 New features and enhancements
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 * `generic.count_domain` used to count values within low and high thresholds.
 * Fire Weather indices now support dask-backed data.
+* Fire Weather Index submodule replicates the R code of cffdrs, including fire season determination and overwintering of the drought_code.
+* New `run_bounds` and `keep_longest_run` utilities in `xclim.indices.run_length`.
 
+Internal changes
+~~~~~~~~~~~~~~~~
+* Small changes in the output of `indices.run_length.rle`.
 
 0.24.0 (2021-03-01)
 -------------------